requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
anaconda-anon-usage=0.4.4=py312hd6b623d_100
anaconda-clean=1.1.1=py312hca03da5_0
anyio=4.3.0=pypi_0
appnope=0.1.4=pypi_0
archspec=0.2.3=pyhd3eb1b0_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.2.0=pypi_0
babel=2.15.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.1.0=pypi_0
boltons=23.0.0=py312hca03da5_0
brotli-python=1.0.9=py312h313beb8_8
bzip2=1.0.8=h80987f9_6
c-ares=1.19.1=h80987f9_0
ca-certificates=2024.3.11=hca03da5_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=py312hca03da5_0
cffi=1.16.0=py312h80987f9_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
conda=24.4.0=py312hca03da5_0
conda-content-trust=0.2.0=py312hca03da5_1
conda-libmamba-solver=24.1.0=pyhd3eb1b0_0
conda-package-handling=2.2.0=py312hca03da5_1
conda-package-streaming=0.9.0=py312hca03da5_0
contourpy=1.2.1=pypi_0
cryptography=42.0.5=py312hd4332d6_1
cssbeautifier=1.15.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
distro=1.9.0=py312hca03da5_0
djlint=1.34.1=pypi_0
editorconfig=0.12.4=pypi_0
executing=2.0.1=pypi_0
expat=2.6.2=h313beb8_0
fastjsonschema=2.19.1=pypi_0
fmt=9.1.0=h48ca7d4_1
fonttools=4.51.0=pypi_0
fqdn=1.5.1=pypi_0
frozendict=2.4.2=py312hca03da5_0
frozenlist=1.4.1=pypi_0
fsspec=2024.5.0=pypi_0
gcsfs=2024.5.0=pypi_0
google-api-core=2.19.0=pypi_0
google-api-python-client=2.130.0=pypi_0
google-auth=2.29.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
google-cloud-core=2.4.1=pypi_0
google-cloud-storage=2.16.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-resumable-media=2.7.0=pypi_0
googleapis-common-protos=1.63.0=pypi_0
h11=0.14.0=pypi_0
html-tag-names=0.1.2=pypi_0
html-void-elements=0.1.0=pypi_0
httpcore=1.0.5=pypi_0
httplib2=0.22.0=pypi_0
httpx=0.27.0=pypi_0
icu=73.1=h313beb8_0
idna=3.7=py312hca03da5_0
ipykernel=6.29.4=pypi_0
ipython=8.24.0=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=pypi_0
jsbeautifier=1.15.1=pypi_0
json5=0.9.25=pypi_0
jsonpatch=1.33=py312hca03da5_1
jsonpointer=2.1=pyhd3eb1b0_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.2.1=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.2=pypi_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=hf3e1bf2_1
libarchive=3.6.2=h62fee54_3
libcurl=8.7.1=h3e2b118_0
libcxx=14.0.6=h848a8c0_0
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libffi=3.4.4=hca03da5_1
libiconv=1.16=h80987f9_3
libmamba=1.5.8=haeffa04_2
libmambapy=1.5.8=py312h1c5506f_2
libnghttp2=1.57.0=h62f6fdd_0
libsolv=0.7.24=h514c7bf_1
libssh2=1.11.0=h3e2b118_0
libxml2=2.10.4=h0b34f26_2
lz4-c=1.9.4=h313beb8_1
markupsafe=2.1.5=pypi_0
matplotlib=3.8.4=pypi_0
matplotlib-inline=0.1.7=pypi_0
menuinst=2.0.2=py312hca03da5_1
mistune=3.0.2=pypi_0
multidict=6.0.5=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.0.13=h1a28f6b_1
overrides=7.7.0=pypi_0
packaging=23.2=py312hca03da5_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pathspec=0.12.1=pypi_0
pcre2=10.42=hb066dcc_1
pexpect=4.9.0=pypi_0
pillow=10.3.0=pypi_0
pip=24.0=py312hca03da5_0
platformdirs=3.10.0=py312hca03da5_0
plotly=5.22.0=pypi_0
pluggy=1.0.0=py312hca03da5_1
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
proto-plus=1.23.0=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pybind11-abi=5=hd3eb1b0_0
pycosat=0.6.6=py312h80987f9_1
pycparser=2.21=pyhd3eb1b0_0
pygments=2.18.0=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py312hca03da5_0
python=3.12.3=h99e199e_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-json-logger=2.0.7=pypi_0
python.app=3=py312h80987f9_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=26.0.3=pypi_0
readline=8.2=h1a28f6b_0
referencing=0.35.1=pypi_0
regex=2023.12.25=pypi_0
reproc=14.2.4=h313beb8_2
reproc-cpp=14.2.4=h313beb8_2
requests=2.31.0=py312hca03da5_1
requests-oauthlib=2.0.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.18.1=pypi_0
rsa=4.9=pypi_0
ruamel.yaml=0.17.21=py312h80987f9_0
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.3=pypi_0
setuptools=69.5.1=py312hca03da5_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.45.3=h80987f9_0
stack-data=0.6.3=pypi_0
tenacity=8.3.0=pypi_0
terminado=0.18.1=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.14=h6ba3021_0
tomlkit=0.12.5=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.2=py312h989b03a_0
traitlets=5.14.3=pypi_0
truststore=0.8.0=py312hca03da5_0
types-python-dateutil=2.9.0.20240316=pypi_0
tzdata=2024.1=pypi_0
uri-template=1.3.0=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=2.1.0=py312hca03da5_1
wcwidth=0.2.13=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.43.0=py312hca03da5_0
xlrd=2.0.1=pypi_0
xz=5.4.6=h80987f9_1
yaml-cpp=0.8.0=h313beb8_1
yarl=1.9.4=pypi_0
zlib=1.2.13=h18a0788_1
zstandard=0.22.0=py312h1a4646a_0
zstd=1.5.5=hd90d995_2
